cap-api-rssfeed/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
channels==4.0.0
channels-redis==4.0.0
pytest-asyncio==0.20.1
pytest-django==4.5.2
Pillow==9.3.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.5
feedparser==6.0.10
python_dateutil==2.8.2
django-cors-headers==4.1.0